refactor: make best practices in prisma folder
