+++ backend/src/health/health.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { HealthController } from './health.controller';
-import { PrismaModule } from '../../prisma/prisma.module';
+import { PrismaModule } from '../prisma/prisma.module';
 
 @Module({
   imports: [PrismaModule],
